Duplicated cabins

+++ src/services/apiCabins.js
@@ -4,9 +4,10 @@ export async function getCabins() {
   const { data, error } = await supabase.from("cabins").select("*");
 
   if (error) {
-    console.log(error);
+    console.error(error);
     throw new Error("Cabins could not be loaded");
   }
+
   return data;
 }
 
@@ -20,34 +21,37 @@ export async function createEditCabin(newCabin, id) {
   const imagePath = hasImagePath
     ? newCabin.image
     : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-  //https://rcjxtdjnwbbtimrkhlgo.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
 
-  //1. Create/edit cabin
+  // 1. Create/edit cabin
   let query = supabase.from("cabins");
 
-  //CREATE
-  if (!id) query.insert([{ ...newCabin, image: imagePath }]);
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-  //EDIT
-  if (id) query.update({ ...newCabin, image: imagePath }).eq("id", id);
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
 
   const { data, error } = await query.select().single();
 
   if (error) {
-    console.log(error);
+    console.error(error);
     throw new Error("Cabin could not be created");
   }
-  //2. Upload image
+
+  // 2. Upload image
+  if (hasImagePath) return data;
 
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  //3. Delete the cabin if error with image
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
-
-    throw new Error("Cabin image could not be uploaded");
+    console.error(storageError);
+    throw new Error(
+      "Cabin image could not be uploaded and the cabin was not created"
+    );
   }
 
   return data;
@@ -57,9 +61,9 @@ export async function deleteCabin(id) {
   const { data, error } = await supabase.from("cabins").delete().eq("id", id);
 
   if (error) {
-    console.log(error);
+    console.error(error);
     throw new Error("Cabin could not be deleted");
   }
+
   return data;
 }
-deleteCabin;
